refactor(dashboard): SlackCard under components/integrations with shared confirm dialog
Post-rebase cleanup for the Integrations hub: the card moves out of
agent-config/ (a leftover from when it lived in AI settings), the
disconnect flow uses the shared useConfirm dialog instead of
window.confirm, and the i18n namespace merge drops the unused
operatorBridges.empty state now that the card always renders.
