+++ packages/dashboard-pages/src/messages/en.json
@@ -1259,18 +1259,6 @@
       "failedDeliveries": "{count} mirror deliveries failed in the last 24 hours."
     }
   },
-  "integrations": {
-    "hero": {
-      "eyebrow": "Integrations",
-      "title": "Connect Munin to <em>the tools you already use</em>",
-      "lede": "Bring conversations into the surfaces your team and customers live in."
-    },
-    "operatorBridges": {
-      "title": "Operator bridges",
-      "blurb": "Triage and reply to conversations from your team's chat tool.",
-      "empty": "Nothing connected yet."
-    }
-  },
   "assistants": {
     "list": {
       "persona": {
